fix: sanitize error messages on empty input
